python/gmail/requirements.txt
google-api-python-client==2.126.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
requests==2.31.0